grpcs/client/requirements.txt
grpcio==1.57.0
grpcio-tools==1.57.0
soundfile==0.12.1
